fix: 修复 32 位平台 go vet 溢出错（qa_perf_test.go 大整数字面量）
将 BenchmarkJSONEncodeResponse 中的 "time": 1700000000000 显式标注为
int64(1700000000000)。该字面量位于 map[string]any 中，默认推断为 int，
在 linux/arm（GOARM=7）等 32 位目标下超出 int32 上限导致 vet 失败。
此问题此前因 CI 矩阵无 32 位目标而未暴露，新增 linux/arm 后触发。

+++ qa_perf_test.go
@@ -207,7 +207,7 @@ func BenchmarkLogWrite(b *testing.B) {
 // BenchmarkJSONEncodeResponse 单个响应 JSON 编码开销。
 func BenchmarkJSONEncodeResponse(b *testing.B) {
 	payload := map[string]any{
-		"status": 200, "time": 1700000000000,
+		"status": 200, "time": int64(1700000000000),
 		"data": map[string]any{"maxPage": 1, "page": 0, "total": 100, "data": []any{}},
 	}
 	var sb strings.Builder
